Safari-611.1.10/LayoutTests/fast/canvas/canvas-gradient-addStop-error-expected.txt

A couple of tests to ensure that adding a stop with an invalid color to a gradient throws exceptions normally.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Ensure exceptions are thrown in normal pages
PASS gradient.addColorStop(1, 'rgb(NaN%, NaN%, NaN%)') threw exception SyntaxError: The string did not match the expected pattern..
PASS gradient.addColorStop(1, 'rgb(NaN%, NaN%, NaN%)') threw exception SyntaxError: The string did not match the expected pattern..
PASS successfullyParsed is true

TEST COMPLETE
